moved base artist to separate file
